requirements.txt
# Core Required Dependencies (Essential for the application to function)
# These packages are required and must be installed
Flask==3.1.0
pandas==2.2.3
numpy==1.26.4
yfinance==0.2.40
scikit-learn==1.5.2
matplotlib==3.9.2
statsmodels==0.14.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2
PyMySQL==1.1.1
python-dotenv==1.0.1
APScheduler==3.10.4

# Testing Dependencies
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==5.0.0

# Optional Dependencies (Application works without these but features are limited)
# Install with: pip install -r requirements-optional.txt for full functionality

# Machine Learning (for LSTM model - requires specific setup)
# tensorflow==2.17.0  # Note: May require special installation on Windows
# keras==3.6.0       # Usually installed with tensorflow

# Social Media APIs (for enhanced sentiment analysis)
# tweepy==4.14.0     # Twitter API access
# textblob==0.18.0   # Text sentiment analysis
# tweet-preprocessor==0.6.0  # Tweet cleaning

# Additional Visualization (optional enhancements)
# seaborn==0.13.2

# Installation Notes:
# 1. Install core dependencies: pip install -r requirements.txt
# 2. For full features: pip install -r requirements-optional.txt
# 3. TensorFlow on Windows may need special installation
# 4. Twitter API requires valid credentials in constants.py
